Re-wrote the gain quantization for split-VQ excitation. Added more bits and quantize one at a time.
git-svn-id: http://svn.xiph.org/trunk/speex@3223 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/mpulse.h
@@ -27,7 +27,7 @@
 typedef struct mpulse_params {
    int     nb_pulse;
    int     nb_tracks;
-   int     gain_coef;
+   float     gain_coef;
    int     track_ind_bits;
 } mpulse_params;
